refactor: introduce Utils.mapFinderOptions
