+++ lib/data-types.js
@@ -676,10 +676,11 @@ UUIDV4.prototype.validate = function(value) {
  * @property VIRTUAL
  * @alias NONE
  */
-var VIRTUAL = function(returnType, fields) {
-  if (!(this instanceof VIRTUAL)) return new VIRTUAL(returnType, fields);
-  
-  this.returnType = returnType;
+var VIRTUAL = function(ReturnType, fields) {
+  if (!(this instanceof VIRTUAL)) return new VIRTUAL(ReturnType, fields);
+  if (typeof ReturnType === 'function') ReturnType = new ReturnType();
+
+  this.returnType = ReturnType;
   this.fields = fields;
 };
 util.inherits(VIRTUAL, ABSTRACT);
