+++ lib/model.js
@@ -538,7 +538,7 @@ validateIncludedElement = function(include, tableNames, options) {
 
   // pseudo include just needed the attribute logic, return
   if (include._pseudo) {
-    return Utils.mapOptionFieldNames(include, include.model);
+    return Utils.mapFinderOptions(include, include.model);
   }
 
   // check if the current Model is actually associated with the passed Model - or it's a pseudo include
@@ -595,7 +595,7 @@ validateIncludedElement = function(include, tableNames, options) {
 
   Model.$injectScope(model.$scope, include);
 
-  include = Utils.mapOptionFieldNames(include, include.model);
+  include = Utils.mapFinderOptions(include, include.model);
 
   if (include.required === undefined) {
     include.required = !!include.where;
@@ -1326,17 +1326,9 @@ Model.prototype.findAll = function(options) {
 
     if (options.attributes === undefined) {
       options.attributes = Object.keys(this.tableAttributes);
-    } else if (this._hasVirtualAttributes) {
-      options.attributes.forEach(function (attribute) {
-        if (this._isVirtualAttribute(attribute) && this.rawAttributes[attribute].type.fields) {
-          options.attributes = options.attributes.concat(this.rawAttributes[attribute].type.fields);
-        }
-      }.bind(this));
-      options.attributes = _.without.apply(this, [options.attributes].concat(this._virtualAttributes));
-      options.attributes = _.unique(options.attributes);
     }
 
-    Utils.mapOptionFieldNames(options, this);
+    Utils.mapFinderOptions(options, this);
 
     options = paranoidClause(this, options);
