+++ lib/utils.js
@@ -76,6 +76,23 @@ var Utils = module.exports = {
     });
   },
 
+  /* Expand and normalize finder options */ 
+  mapFinderOptions: function(options, Model) {
+    if (Model._hasVirtualAttributes) {
+      options.attributes.forEach(function (attribute) {
+        if (Model._isVirtualAttribute(attribute) && Model.rawAttributes[attribute].type.fields) {
+          options.attributes = options.attributes.concat(Model.rawAttributes[attribute].type.fields);
+        }
+      }.bind(Model));
+      options.attributes = _.without.apply(_, [options.attributes].concat(Model._virtualAttributes));
+      options.attributes = _.unique(options.attributes);
+    }
+
+    Utils.mapOptionFieldNames(options, Model);
+
+    return options;
+  },
+
   /* Used to map field names in attributes and where conditions */
   mapOptionFieldNames: function(options, Model) {
     if (options.attributes) {
